+++ Sources/StreamChat/Database/DTOs/ChannelDTO.swift
@@ -373,18 +373,6 @@ extension ChannelDTO {
         request.fetchBatchSize = query.pagination.pageSize
         return request
     }
-
-    static func channelsFetchRequest(notLinkedTo query: ChannelListQuery) -> NSFetchRequest<ChannelDTO> {
-        let request = NSFetchRequest<ChannelDTO>(entityName: ChannelDTO.entityName)
-        request.sortDescriptors = [ChannelListSortingKey.defaultSortDescriptor]
-        // Channels which are not linked to this query
-        request.predicate = NSCompoundPredicate(
-            notPredicateWithSubpredicate: NSPredicate(
-                format: "ANY queries.filterHash == %@", query.filter.filterHash
-            )
-        )
-        return request
-    }
 }
 
 extension ChannelDTO {