+++ StreamChatUITestsAppUITests/Tests/MessageList_Tests.swift
@@ -469,6 +469,8 @@ extension MessageList_Tests {
     }
     
     func test_scrollToBottom_unreadCount() throws {
+        throw XCTSkip("https://github.com/GetStream/ios-issues-tracking/issues/491")
+        
         linkToScenario(withId: 1669)
         
         try XCTSkipIf(ProcessInfo().operatingSystemVersion.majorVersion == 12,