+++ Tests/StreamChatTests/Controllers/ChannelListController/ChannelListController_Tests.swift
@@ -277,242 +277,6 @@ final class ChannelListController_Tests: XCTestCase {
         AssertAsync.willBeEqual(controller.channels.map(\.cid), [cid])
     }
 
-    func test_newChannel_callsListHook_whenSynchronized() throws {
-        setUpChatClientWithoutAutoFiltering()
-
-        // Simulate `synchronize` call and catch the completion
-        var synchronized = false
-        controller.synchronize { _ in synchronized = true }
-
-        // Add the channel to the DB
-        let cid: ChannelId = .unique
-        let channelPayload = dummyPayload(with: cid, members: [.dummy(user: .dummy(userId: memberId))])
-        var channel: ChatChannel!
-        try database.writeSynchronously { session in
-            let dto = try session.saveChannel(payload: channelPayload, query: self.query, cache: nil)
-            channel = try dto.asModel()
-        }
-
-        // Simulate successful response from backend
-        env.channelListUpdater?.update_completion?(.success([channel]))
-
-        AssertAsync {
-            // Assert synchronized completion is invoked
-            Assert.willBeTrue(synchronized)
-            // Assert the resulting value is updated
-            Assert.willBeEqual(self.controller.channels.map(\.cid), [cid])
-        }
-
-        let newCid: ChannelId = .unique
-
-        // Create and assign delegate
-        let delegate = TestLinkDelegate(shouldListNewChannel: { channel in
-            channel.cid != newCid
-        }, shouldListUpdatedChannel: { _ in
-            false
-        })
-        controller.delegate = delegate
-
-        // Insert a new channel to DB
-        try database.writeSynchronously { session in
-            try session.saveChannel(payload: .dummy(cid: newCid, members: [.dummy(user: .dummy(userId: self.memberId))]), query: nil, cache: nil)
-        }
-
-        // Assert the resulting value is not inserted
-        AssertAsync.willBeEqual(controller.channels.map(\.cid), [cid])
-
-        // Insert a new channel to DB
-        let insertedCid = ChannelId.unique
-        try database.writeSynchronously { session in
-            try session.saveChannel(payload: .dummy(cid: insertedCid, members: [.dummy(user: .dummy(userId: self.memberId))]), query: nil, cache: nil)
-        }
-
-        // Assert the resulting value is inserted
-        AssertAsync.willBeEqual(controller.channels.map(\.cid.rawValue).sorted(), [cid.rawValue, insertedCid.rawValue].sorted())
-    }
-
-    func test_updatedChannel_callsLinkHook_whenSynchronized() throws {
-        setUpChatClientWithoutAutoFiltering()
-
-        // Simulate `synchronize` call and catch the completion
-        var synchronized = false
-        controller.synchronize { _ in synchronized = true }
-
-        // Add the channel to the DB
-        let cid: ChannelId = .unique
-        let channelPayload = dummyPayload(with: cid, members: [.dummy(user: .dummy(userId: memberId))])
-        var channel: ChatChannel!
-        try database.writeSynchronously { session in
-            let dto = try session.saveChannel(payload: channelPayload, query: self.query, cache: nil)
-            channel = try dto.asModel()
-        }
-
-        // Simulate successful response from backend
-        env.channelListUpdater?.update_completion?(.success([channel]))
-
-        AssertAsync {
-            // Assert synchronized completion is invoked
-            Assert.willBeTrue(synchronized)
-            // Assert the resulting value is updated
-            Assert.willBeEqual(self.controller.channels.map(\.cid), [cid])
-        }
-
-        let shouldBeInsertedCid: ChannelId = .unique
-        let shouldBeExcludedCid: ChannelId = .unique
-
-        // Create and assign delegate
-        let delegate = TestLinkDelegate(shouldListNewChannel: { _ in
-            false
-        }, shouldListUpdatedChannel: { channel in
-            channel.cid == shouldBeInsertedCid
-        })
-        controller.delegate = delegate
-
-        // Insert 2 channels to cid
-        try database.writeSynchronously { session in
-            try session.saveChannel(payload: .dummy(cid: shouldBeInsertedCid, members: [.dummy(user: .dummy(userId: self.memberId))]), query: nil, cache: nil)
-            try session.saveChannel(payload: .dummy(cid: shouldBeExcludedCid, members: [.dummy(user: .dummy(userId: self.memberId))]), query: nil, cache: nil)
-        }
-
-        // Assert that 2 new channels are not linked
-        AssertAsync.willBeEqual(controller.channels.map(\.cid), [cid])
-
-        // Update `shouldBeExcludedCid`
-        try database.writeSynchronously { session in
-            let dto = try XCTUnwrap(session.channel(cid: shouldBeExcludedCid))
-            dto.updatedAt = .unique
-        }
-
-        // Assert that updated channel is not linked
-        AssertAsync.willBeEqual(controller.channels.map(\.cid), [cid])
-
-        // Update `shouldBeInsertedCid`
-        try database.writeSynchronously { session in
-            let dto = try XCTUnwrap(session.channel(cid: shouldBeInsertedCid))
-            dto.updatedAt = .unique
-        }
-
-        // Assert that updated channel is linked
-        AssertAsync.willBeEqual(
-            controller.channels.map(\.cid.rawValue).sorted(),
-            [cid.rawValue, shouldBeInsertedCid.rawValue].sorted()
-        )
-    }
-
-    func test_updatedChannel_callsUnlinkHook_whenSynchronized() throws {
-        setUpChatClientWithoutAutoFiltering()
-
-        // Simulate `synchronize` call and catch the completion
-        var synchronized = false
-        controller.synchronize { _ in synchronized = true }
-
-        // Add the channel to the DB
-        let cid: ChannelId = .unique
-        let channelPayload = dummyPayload(with: cid, members: [.dummy(user: .dummy(userId: memberId))])
-        var channel: ChatChannel!
-        try database.writeSynchronously { session in
-            let dto = try session.saveChannel(payload: channelPayload, query: self.query, cache: nil)
-            channel = try dto.asModel()
-        }
-
-        // Simulate successful response from backend
-        env.channelListUpdater?.update_completion?(.success([channel]))
-
-        AssertAsync {
-            // Assert synchronized completion is invoked
-            Assert.willBeTrue(synchronized)
-            // Assert the resulting value is updated
-            Assert.willBeEqual(self.controller.channels.map(\.cid), [cid])
-        }
-
-        // Create and assign delegate
-        let delegate = TestLinkDelegate(
-            shouldListNewChannel: { _ in false },
-            shouldListUpdatedChannel: { channel in
-                channel.cid != cid
-            }
-        )
-        controller.delegate = delegate
-
-        // Update linked channel
-        try database.writeSynchronously { session in
-            let channelDTO = session.channel(cid: cid)
-            channelDTO?.updatedAt = .unique
-        }
-
-        // Assert that new channel is unlinked
-        AssertAsync.willBeEqual(controller.channels.map(\.cid), [])
-    }
-
-    func test_unlinkedChannels_doNotTriggerHooks_whenNotSynchronized() throws {
-        // Create and assign delegate, catch
-        var delegateCalled = false
-        let delegate = TestLinkDelegate(shouldListNewChannel: { _ in
-            delegateCalled = true
-            return false
-        }, shouldListUpdatedChannel: { _ in
-            delegateCalled = true
-            return false
-        })
-        controller.delegate = delegate
-
-        // Save a channel not-linked to the current query
-        let cid: ChannelId = .unique
-        try database.writeSynchronously { session in
-            try session.saveChannel(payload: self.dummyPayload(with: cid), query: nil, cache: nil)
-        }
-
-        AssertAsync {
-            // Assert hooks are not called
-            Assert.staysFalse(delegateCalled)
-            // Assert channels stay empty
-            Assert.willBeEqual(self.controller.channels.map(\.cid), [])
-        }
-
-        // Update a channel not linked to the current query
-        try database.writeSynchronously { session in
-            let dto = try XCTUnwrap(session.channel(cid: cid))
-            dto.updatedAt = .unique
-        }
-
-        AssertAsync {
-            // Assert hooks are not called
-            Assert.staysFalse(delegateCalled)
-            // Assert channels stay empty
-            Assert.willBeEqual(self.controller.channels.map(\.cid), [])
-        }
-    }
-
-    func test_linkedChannels_doesTriggerUnlinkHook_whenNotSynchronized() throws {
-        setUpChatClientWithoutAutoFiltering()
-
-        // Save a channel linked to the current query
-        let cid: ChannelId = .unique
-        try database.writeSynchronously { session in
-            try session.saveChannel(payload: self.dummyPayload(with: cid, members: [.dummy(user: .dummy(userId: self.memberId))]), query: self.query, cache: nil)
-        }
-
-        // Assert channel is linked
-        AssertAsync.willBeEqual(controller.channels.map(\.cid), [cid])
-
-        // Create and assign delegate excluding a channel from a query
-        let delegate = TestLinkDelegate(shouldListNewChannel: { _ in
-            true
-        }, shouldListUpdatedChannel: { channel in
-            channel.cid != cid
-        })
-        controller.delegate = delegate
-
-        // Update a channel linked to the current query
-        try database.writeSynchronously { session in
-            let dto = try XCTUnwrap(session.channel(cid: cid))
-            dto.updatedAt = .unique
-        }
-
-        // Assert linked channel is unlisted
-        AssertAsync.willBeEqual(controller.channels.map(\.cid), [])
-    }
-
     func test_hiddenChannel_isExcluded_whenFilterDoesntContainHiddenKey() throws {
         // Add 2 channels to the DB
         let cid: ChannelId = .unique
@@ -553,127 +317,97 @@ final class ChannelListController_Tests: XCTestCase {
         XCTAssertEqual(controller.channels.first?.isHidden, true)
     }
 
-    // MARK: - Change propagation tests with filter block
+    // MARK: - Linking and Unlinking Channels when channels are updated/inserted
 
-    func test_addingANewChannelThatMatchesFilter_shouldLinkItToQuery_shouldWatchIt() throws {
-        var filterCalls = 0
-        prepareControllerWithOwnerFilter(userId: memberId, onFilterCall: { filterCalls += 1 })
+    func test_didReceiveEvent_whenNotificationAddedToChannelEvent_shouldLinkChannelToQuery() {
+        let event = makeAddedChannelEvent(with: .mock(cid: .unique))
 
-        // Simulate `synchronize` call
-        let expectation = self.expectation(description: "Synchronize completion")
-        controller.synchronize { _ in expectation.fulfill() }
-        let originalCid: ChannelId = .unique
-        try addOrUpdateChannel(cid: originalCid, ownerId: memberId, query: query)
+        controller.eventsController(controller.eventsController, didReceiveEvent: event)
 
-        // Simulate successful response from backend
-        env.channelListUpdater?.update_completion?(.success([]))
-        waitForExpectations(timeout: defaultTimeout, handler: nil)
-        XCTAssertEqual(controller.channels.count, 1)
-        XCTAssertEqual(filterCalls, 0)
+        env.channelListUpdater?.link_completion?(nil)
 
-        let ownedCid: ChannelId = .unique
-        let unownedCid: ChannelId = .unique
-        try [ownedCid, unownedCid].forEach { cid in
-            try addOrUpdateChannel(cid: cid, ownerId: cid == ownedCid ? memberId : "other", query: nil)
-        }
-
-        // The original channel + the owned that was just added
-        AssertAsync.willBeEqual(controller.channels.count, 2)
-        XCTAssertTrue(controller.channels.contains { $0.cid == originalCid })
-        XCTAssertTrue(controller.channels.contains { $0.cid == ownedCid })
-        XCTAssertFalse(controller.channels.contains { $0.cid == unownedCid })
-
-        // Two channels were added to the DB, both were evaluated, only one was added to the list
-        XCTAssertEqual(filterCalls, 2)
-        // Watches channel that was added to the list
-        XCTAssertEqual(env.channelListUpdater?.startWatchingChannels_cids, [ownedCid])
+        XCTAssertEqual(env.channelListUpdater?.link_callCount, 1)
+        XCTAssertEqual(env.channelListUpdater?.startWatchingChannels_callCount, 1)
     }
 
-    func test_updatingAChannelThatIsLinkedToTheQuery_shouldUnlinkIt() throws {
-        var filterCalls = 0
-        prepareControllerWithOwnerFilter(userId: memberId, onFilterCall: { filterCalls += 1 })
+    func test_didReceiveEvent_whenMessageNewEvent_shouldLinkChannelToQuery() {
+        let event = makeMessageNewEvent(with: .mock(cid: .unique))
 
-        // Simulate `synchronize` call
-        let expectation = self.expectation(description: "Synchronize completion")
-        controller.synchronize { _ in expectation.fulfill() }
-        let originalCid: ChannelId = .unique
-        try addOrUpdateChannel(cid: originalCid, ownerId: memberId, query: query)
+        controller.eventsController(controller.eventsController, didReceiveEvent: event)
 
-        // Simulate successful response from backend
-        env.channelListUpdater?.update_completion?(.success([]))
-        waitForExpectations(timeout: defaultTimeout, handler: nil)
-        XCTAssertEqual(controller.channels.count, 1)
-        XCTAssertEqual(filterCalls, 0)
+        env.channelListUpdater?.link_completion?(nil)
 
-        // Simulate update of the channel linked to query
-        try addOrUpdateChannel(cid: originalCid, ownerId: "something else", query: nil)
-
-        // The original channel + the owned that was just added
-        AssertAsync.willBeEqual(controller.channels.count, 0)
-        XCTAssertFalse(controller.channels.contains { $0.cid == originalCid })
-
-        // The updated channel was evaluated once to be unlinked, and then by the unlinked query listener
-        XCTAssertEqual(filterCalls, 2)
+        XCTAssertEqual(env.channelListUpdater?.link_callCount, 1)
+        XCTAssertEqual(env.channelListUpdater?.startWatchingChannels_callCount, 1)
     }
 
-    func test_updatingAChannelThatIsNotLinkedToTheQuery_shouldLinkIt() throws {
-        var filterCalls = 0
-        prepareControllerWithOwnerFilter(userId: memberId, onFilterCall: { filterCalls += 1 })
+    func test_didReceiveEvent_whenNotificationMessageNewEvent_shouldLinkChannelToQuery() {
+        let event = makeNotificationMessageNewEvent(with: .mock(cid: .unique))
 
-        // Simulate `synchronize` call
-        let expectation = self.expectation(description: "Synchronize completion")
-        controller.synchronize { _ in expectation.fulfill() }
-        let originalCid: ChannelId = .unique
-        try addOrUpdateChannel(cid: originalCid, ownerId: memberId, query: query)
+        controller.eventsController(controller.eventsController, didReceiveEvent: event)
 
-        // Simulate successful response from backend
-        env.channelListUpdater?.update_completion?(.success([]))
-        waitForExpectations(timeout: defaultTimeout, handler: nil)
-        XCTAssertEqual(controller.channels.count, 1)
-        XCTAssertEqual(filterCalls, 0)
+        env.channelListUpdater?.link_completion?(nil)
 
-        let originallyUnownedCid: ChannelId = .unique
-        try addOrUpdateChannel(cid: originallyUnownedCid, ownerId: "other", query: nil)
-
-        // The added channel should be evaluated, but not added
-        AssertAsync.willBeEqual(filterCalls, 1)
-        XCTAssertTrue(controller.channels.contains { $0.cid == originalCid })
-        XCTAssertFalse(controller.channels.contains { $0.cid == originallyUnownedCid })
-
-        // Simulate update on the channel not matching the query, to make it match it
-        try addOrUpdateChannel(cid: originallyUnownedCid, ownerId: memberId, query: nil)
-
-        // The updated channel should be evaluated, and added
-        AssertAsync.willBeEqual(controller.channels.count, 2)
-        XCTAssertEqual(filterCalls, 2)
-        XCTAssertTrue(controller.channels.contains { $0.cid == originalCid })
-        XCTAssertTrue(controller.channels.contains { $0.cid == originallyUnownedCid })
+        XCTAssertEqual(env.channelListUpdater?.link_callCount, 1)
+        XCTAssertEqual(env.channelListUpdater?.startWatchingChannels_callCount, 1)
     }
 
-    private func addOrUpdateChannel(cid: ChannelId, ownerId: String, query: ChannelListQuery?) throws {
-        try database.writeSynchronously { session in
-            let payload = self.dummyPayload(with: cid, members: [.dummy(user: .dummy(userId: self.memberId))], channelExtraData: ["owner_id": .string(ownerId)])
-            try session.saveChannel(payload: payload, query: query, cache: nil)
-        }
-    }
-
-    private func prepareControllerWithOwnerFilter(userId: UserId, onFilterCall: @escaping () -> Void) {
+    func test_didReceiveEvent_whenFilterMatches_shouldLinkChannelToQuery() {
         let filter: (ChatChannel) -> Bool = { channel in
-            onFilterCall()
-            guard case let .string(owner) = channel.extraData["owner_id"] else { return false }
-            return owner == userId
+            channel.memberCount == 4
         }
+        setupControllerWithFilter(filter)
 
-        // Prepare controller
+        let event = makeAddedChannelEvent(with: .mock(cid: .unique, memberCount: 4))
 
-        client.authenticationRepository.setMockToken()
-        query = .init(filter: .and(
-            [
-                .containMembers(userIds: [userId]),
-                .notEqual(.init(rawValue: "owner_id"), to: userId)
-            ]
-        ))
-        controller = ChatChannelListController(query: query, client: client, filter: filter, environment: env.environment)
+        controller.eventsController(controller.eventsController, didReceiveEvent: event)
+
+        env.channelListUpdater?.link_completion?(nil)
+
+        XCTAssertEqual(env.channelListUpdater?.link_callCount, 1)
+        XCTAssertEqual(env.channelListUpdater?.startWatchingChannels_callCount, 1)
+    }
+
+    func test_didReceiveEvent_whenFilterDoesNotMatch_shouldNotLinkChannelToQuery() {
+        let filter: (ChatChannel) -> Bool = { channel in
+            channel.memberCount == 1
+        }
+        setupControllerWithFilter(filter)
+
+        let event = makeAddedChannelEvent(with: .mock(cid: .unique, memberCount: 4))
+
+        controller.eventsController(controller.eventsController, didReceiveEvent: event)
+
+        env.channelListUpdater?.link_completion?(nil)
+
+        XCTAssertEqual(env.channelListUpdater?.link_callCount, 0)
+        XCTAssertEqual(env.channelListUpdater?.startWatchingChannels_callCount, 0)
+    }
+
+    func test_didReceiveEvent_whenChannelUpdatedEvent_whenFilterDoesNotMatch_shouldUnlinkChannelFromQuery() {
+        let filter: (ChatChannel) -> Bool = { channel in
+            channel.memberCount == 1
+        }
+        setupControllerWithFilter(filter)
+
+        let event = makeChannelUpdatedEvent(with: .mock(cid: .unique, memberCount: 4))
+
+        controller.eventsController(controller.eventsController, didReceiveEvent: event)
+
+        XCTAssertEqual(env.channelListUpdater?.unlink_callCount, 1)
+    }
+
+    func test_didReceiveEvent_whenChannelUpdatedEvent__whenFilterMatches_shouldNotUnlinkChannelFromQuery() {
+        let filter: (ChatChannel) -> Bool = { channel in
+            channel.memberCount == 4
+        }
+        setupControllerWithFilter(filter)
+
+        let event = makeChannelUpdatedEvent(with: .mock(cid: .unique, memberCount: 4))
+
+        controller.eventsController(controller.eventsController, didReceiveEvent: event)
+
+        XCTAssertEqual(env.channelListUpdater?.unlink_callCount, 0)
     }
 
     // MARK: - Change propagation tests with auto-filtering
@@ -1750,6 +1484,55 @@ final class ChannelListController_Tests: XCTestCase {
     }
 
     // MARK: - Private Helpers
+
+    private func makeAddedChannelEvent(with channel: ChatChannel) -> NotificationAddedToChannelEvent {
+        NotificationAddedToChannelEvent(
+            channel: channel,
+            unreadCount: nil,
+            member: .mock(id: .unique),
+            createdAt: .unique
+        )
+    }
+
+    private func makeMessageNewEvent(with channel: ChatChannel) -> MessageNewEvent {
+        MessageNewEvent(
+            user: .unique,
+            message: .unique,
+            channel: channel,
+            createdAt: .unique,
+            watcherCount: nil,
+            unreadCount: nil
+        )
+    }
+
+    private func makeNotificationMessageNewEvent(with channel: ChatChannel) -> NotificationMessageNewEvent {
+        NotificationMessageNewEvent(
+            channel: channel,
+            message: .unique,
+            createdAt: .unique,
+            unreadCount: nil
+        )
+    }
+
+    private func makeChannelUpdatedEvent(with channel: ChatChannel) -> ChannelUpdatedEvent {
+        ChannelUpdatedEvent(
+            channel: channel,
+            user: .unique,
+            message: .unique,
+            createdAt: .unique
+        )
+    }
+
+    private func setupControllerWithFilter(_ filter: @escaping (ChatChannel) -> Bool) {
+        // Prepare controller
+        controller = ChatChannelListController(
+            query: query,
+            client: client,
+            filter: filter,
+            environment: env.environment
+        )
+        controller.synchronize()
+    }
 
     private func setUpChatClientWithoutAutoFiltering() {
         var config = ChatClientConfig(apiKey: .init(.unique))